django_project/requirement.txt
certifi==2019.6.16
Django==2.2.1
django-crispy-forms==1.7.2
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
wincertstore==0.2
